challenges/web/pdftex/requirements.txt
Flask
Flask-MySQLdb
Flask-WTF
Flask-Login
Werkzeug